fixing unit test

+++ onebusaway-nyc-tdm-adapters/src/main/java/org/onebusaway/nyc/transit_data_manager/siri/NycSiriService.java
@@ -233,7 +233,8 @@ public abstract class NycSiriService {
     }
 
     // clean up any orphan records
-    getServiceAlertsPersister().deleteOrphans();
+    if (getServiceAlertsPersister() != null) // protect existing unit tests
+      getServiceAlertsPersister().deleteOrphans();
 
   }
